Events listener using Predicate selectors instead of cssSelectors
This is a work around of the performance problem caused by closest->is(cssSelector) so you can implement your own Predicate filter and skip the closest call. This problem get worst when the target event is a child of a element with lot of children, because the current implementation uses $(e.getParentNode(),cssSelector).first()==e to check if e satisfies cssSelector.